@@ -35,17 +35,15 @@
     "babel-plugin-transform-es2015-modules-amd": "^6.24.1",
     "babel-preset-env": "^1.7.0",
     "babel-preset-stage-2": "^6.24.1",
-    "eslint": "^4.18.0",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-prettier": "^2.6.0",
+    "eslint": "^4.19.1",
+    "eslint-plugin-import": "^2.17.3",
     "flow-bin": "0.65.0",
     "gulp": "^3.9.1",
     "gulp-add-src": "^1.0.0",
     "gulp-babel": "^7.0.1",
     "gulp-clean": "^0.4.0",
     "gulp-concat": "^2.6.1",
-    "gulp-eslint": "^4.0.2",
+    "gulp-eslint": "^5.0.0",
     "gulp-gzip": "^1.4.2",
     "gulp-header": "^2.0.5",
     "gulp-jsonlint": "^1.2.1",
@@ -58,7 +56,6 @@
     "mongodb-memory-server": "^1.7.0",
     "nodemon": "^1.18.10",
     "path": "^0.12.7",
-    "prettier": "^1.16.4",
     "request": "^2.83.0",
     "run-sequence": "^2.2.1"
   },
@@ -106,12 +103,14 @@
     "lodash": "^4.17.5",
     "mixitup": "^3.2.2",
     "moment": "^2.21.0",
+    "mongodb": "^3.2.7",
     "mongoose": "^5.4.11",
     "mongoose-uuid2": "^2.1.0",
     "multer": "^1.4.0",
     "neat-csv": "^3.0.0",
     "node-uuid": "^1.4.8",
     "subscriptions-transport-ws": "^0.9.7",
+    "through2": "^3.0.1",
     "type-detect": "^4.0.8",
     "useragent": "^2.3.0",
     "utf-8-validate": "^4.0.1",
