@@ -2,7 +2,7 @@ import Validator from 'ajv';
 import _ from 'lodash';
 import refParser from 'json-schema-ref-parser';
 
-const { find, readFile } = require('./util/fs');
+import { find } from './util/fs';
 
 
 export const loadValidator = function(schemas) {
@@ -29,7 +29,8 @@ export const loadValidator = function(schemas) {
 			function validate(schema, value) {
 				try {
 					validator.validate(schema, value);
-				} catch(err) {
+				}
+				catch (err) {
 					return Promise.reject(err);
 				}
 
