@@ -1,6 +1,5 @@
 'use strict';
 
-import config from 'config';
 import express from 'express';
 import httpErrors from 'http-errors';
 
@@ -10,7 +9,7 @@ let router = express.Router();
 let accessToken = router.route('/');
 
 
-accessToken.post(async function(req, res, next) {
+accessToken.post(async function(req, res) {
 	let args = {
 		client_id: req.body.client_id,
 		client_secret: req.body.client_secret,
@@ -48,4 +47,4 @@ accessToken.post(async function(req, res, next) {
 	}
 });
 
-module.exports = router;
+export default router;
