@@ -26,4 +26,4 @@ logout.get(function(req, res, next) {
 		});
 });
 
-module.exports = router;
+export default logout;
