@@ -1,8 +1,7 @@
-import config from 'config';
 import moment from 'moment';
 
-import {SessionTC} from '../../schema/models/sessions';
-import {UserTC} from '../../schema/models/users';
+import { SessionTC } from '../../schema/models/sessions';
+import { UserTC } from '../../schema/models/users';
 
 
 export default async function(req, res, next) {
@@ -61,4 +60,4 @@ export default async function(req, res, next) {
 	}
 
 	next();
-};
+}
