@@ -46,7 +46,8 @@ async function validate(req, res, next) {
 			}
 
 			next();
-		} catch (err) {
+		}
+		catch (err) {
 			console.log(err);
 
 			next(err);
