@@ -1,6 +1,8 @@
+/* global env */
+
 'use strict';
 
-import {UserTC} from '../../schema/models/users';
+import { UserTC } from '../../schema/models/users';
 
 
 let re_token = /^Key ([a-fA-F0-9]{32})$/;
@@ -32,4 +34,4 @@ export default async function(req, res, next) {
 	}
 
 	next();
-};
+}
