@@ -2,29 +2,29 @@ import _ from 'lodash';
 import httpErrors from 'http-errors';
 
 
-export default function(Resolver, scopes, resolvers: { [name: string]: Resolver }) {
-  const secureResolvers = {};
-  let scopeArray = Array.isArray(scopes) ? scopes : [scopes];
-  
-  Object.keys(resolvers).forEach(k => {
-    secureResolvers[k] = resolvers[k].wrapResolve(next => rp => {
-      let approved = false;
+export default function(Resolver, scopes, resolvers) {
+	const secureResolvers = {};
+	let scopeArray = Array.isArray(scopes) ? scopes : [scopes];
 
-      if (rp.context.req.scopes) {
-          _.each(scopeArray, function(scope) {
-            if (rp.context.req.scopes.indexOf(scope) >= 0) {
-              approved = true;
-            }
-          });
+	Object.keys(resolvers).forEach(k => {
+		secureResolvers[k] = resolvers[k].wrapResolve(next => rp => {
+			let approved = false;
 
-          if (approved === false) {
-	          throw new httpErrors(401, 'Token does not have the appropriate scope');
-          }
-      }
+			if (rp.context.req.scopes) {
+				_.each(scopeArray, function(scope) {
+					if (rp.context.req.scopes.indexOf(scope) >= 0) {
+						approved = true;
+					}
+				});
 
-    return next(rp);
-    });
-  });
-  
-  return secureResolvers;
+				if (approved === false) {
+					throw new httpErrors(401, 'Token does not have the appropriate scope');
+				}
+			}
+
+			return next(rp);
+		});
+	});
+
+	return secureResolvers;
 }
