@@ -1,10 +1,9 @@
 'use strict';
 
-import httpErrors from 'http-errors';
 import moment from 'moment';
 
-import {OAuthTokenTC} from "../../schema/models/oauth-tokens";
-import {UserTC} from '../../schema/models/users';
+import { OAuthTokenTC } from "../../schema/models/oauth-tokens";
+import { UserTC } from '../../schema/models/users';
 
 
 let re_token = /^Bearer ([a-fA-F0-9]{64})$/;
@@ -47,4 +46,4 @@ export default async function(req, res, next) {
 	}
 
 	next();
-};
+}
