@@ -1,8 +1,7 @@
-import config from 'config';
 import moment from 'moment';
 
-import {SessionTC} from '../../schema/models/sessions';
-import {UserTC} from '../../schema/models/users';
+import { SessionTC } from '../../schema/models/sessions';
+import { UserTC } from '../../schema/models/users';
 
 
 export default async function(cookie) {
@@ -46,4 +45,4 @@ export default async function(cookie) {
 
 		return userResult || null;
 	}
-};
+}
