@@ -1,5 +1,6 @@
+/* global env */
+
 import _ from 'lodash';
-import httpErrors from 'http-errors';
 import moment from 'moment';
 
 import uuid from './uuid';
