@@ -1,9 +1,9 @@
+/* global env */
+
 import _ from 'lodash';
 import httpErrors from 'http-errors';
 
-import uuid from '../../lib/util/uuid';
-
-import { ContactTC} from "../../schema/models/contacts";
+import { ContactTC } from "../../schema/models/contacts";
 import { ConnectionTC } from "../../schema/models/connections";
 import { ContentTC } from "../../schema/models/content";
 import { EventTC } from "../../schema/models/events";
@@ -42,7 +42,8 @@ export default async function(connectionId, userId) {
 
 		try {
 			await bitscoop.deleteConnection(bitscoopConnectionId);
-		} catch(err) {
+		}
+		catch (err) {
 			console.log('Bad BitScoop Connection ID for ' + connection._id.toString('hex'));
 		}
 	}
