@@ -1,3 +1,5 @@
+/* global env */
+
 import _ from 'lodash';
 import config from 'config';
 import express from 'express';
@@ -11,7 +13,7 @@ import createSelfContact from '../../util/create-self-contact'
 import { AssociationSessionTC } from '../../../schema/models/association-sessions';
 import { ConnectionTC } from '../../../schema/models/connections';
 import { ProviderTC } from '../../../schema/models/providers';
-import {SessionTC} from "../../../schema/models/sessions";
+import { SessionTC } from "../../../schema/models/sessions";
 import { UserTC } from '../../../schema/models/users';
 
 
@@ -22,7 +24,7 @@ let complete = router.route('/');
 
 async function completeHandler(req, res, next) {
 	await env.validate('#/types/uuid4', req.query.connection_id)
-		.catch(function(err) {
+		.catch(function() {
 			throw new Error('Connection ID must be a 32-character UUID with no dashes.')
 		});
 
@@ -61,7 +63,7 @@ async function completeHandler(req, res, next) {
 }
 
 
-async function completeLogin(req, res, next) {
+async function completeLogin(req, res) {
 	let bitscoop = env.bitscoop;
 
 	if (req.user == null) {
@@ -143,7 +145,7 @@ async function completeLogin(req, res, next) {
 							user_id_string: user._id.toString('hex')
 						}
 					}
-				})
+				});
 			}
 
 			let session = await CreateSession(req, user, {
@@ -176,7 +178,8 @@ async function completeLogin(req, res, next) {
 			});
 
 			return user;
-		} catch(err) {
+		}
+		catch (err) {
 			console.log(err);
 
 			res.redirect('/')
@@ -214,11 +217,12 @@ async function completeLogin(req, res, next) {
 							user_id_string: req.user._id.toString('hex')
 						}
 					}
-				})
+				});
 			}
 
 			return req.user;
-		} catch(err) {
+		}
+		catch (err) {
 			console.log(err);
 
 			res.redirect('/providers');
@@ -312,7 +316,8 @@ async function completeConnection(req, res, next, user) {
 				}
 			});
 		}
-	} catch(err) {
+	}
+	catch (err) {
 		throw err;
 	}
 }
