@@ -1,6 +1,5 @@
 'use strict';
 
-import config from 'config';
 import express from 'express';
 
 import csrf from '../middleware/csrf';
@@ -8,10 +7,8 @@ import csrf from '../middleware/csrf';
 let router = express.Router();
 let csrfRoute = router.route('/');
 
-let domain = config.domain;
 
-
-async function csrfHandler(req, res, next) {
+async function csrfHandler(req, res) {
 	res.json({
 		csrf_token: res.context.csrf_token
 	});
