@@ -9,7 +9,7 @@ import { EventTC } from '../../../schema/models/events';
 import csrf from '../../middleware/csrf';
 import multipartParser from '../../middleware/multipart-parser';
 import uuid from '../../util/uuid';
-import {ProviderTC} from '../../../schema/models/providers';
+import { ProviderTC } from '../../../schema/models/providers';
 
 let router = express.Router();
 
@@ -20,14 +20,15 @@ let financialProviderId = 'd2b24c35ffbb47d694ec7a2951247c88';
 let tagRegex = /#[^#\s]+/g;
 
 
-async function uploadHandler(req, res, next) {
+async function uploadHandler(req, res) {
 	try {
 		let body;
 		res.type('json');
 
 		try {
 			body = await csvParser(req.file.buffer);
-		} catch(err) {
+		}
+		catch (err) {
 			return httpErrors(400, 'Invalid CSV file');
 		}
 
@@ -153,7 +154,8 @@ async function uploadHandler(req, res, next) {
 		});
 
 		res.sendStatus(204);
-	} catch(err) {
+	}
+	catch (err) {
 		console.log(err);
 
 		res.sendStatus(400);
