@@ -18,7 +18,7 @@ router.use('/financials', financials);
 router.use('/locations', locations);
 router.use('/logout', logout);
 
-router.use('/health', function(req, res, next) {
+router.use('/health', function(req, res) {
 	res.sendStatus(204);
 });
 
