@@ -30,13 +30,14 @@ let router = express.Router();
 let uploadFile = router.route('/');
 
 
-async function uploadHandler(req, res, next) {
+async function uploadHandler(req, res) {
 	let body;
 	res.type('json');
 
 	try {
 		body = JSON.parse(req.file.buffer);
-	} catch(err) {
+	}
+	catch (err) {
 		return httpErrors(400, 'Invalid JSON');
 	}
 
@@ -52,15 +53,15 @@ async function uploadHandler(req, res, next) {
 
 	new Promise(function(resolve, reject) {
 		s3.upload({
-			Bucket : config.aws.s3.locations.bucket_name,
+			Bucket: config.aws.s3.locations.bucket_name,
 			Key: req.user._id.toString('hex') + '/' + startTime.unix() + '.json',
 			Body: req.file.buffer
-		}, async function(err, data) {
+		}, async function(err) {
 			if (err) {
 				reject(err);
 			}
 			else {
-				let endTime = moment().utc();
+				// let endTime = moment().utc();
 
 				try {
 					await mongoose.connection.db.collection('location_files').insert({
@@ -70,15 +71,16 @@ async function uploadHandler(req, res, next) {
 						status: 'ready',
 						user_id: req.user._id
 					});
-				} catch(err) {
-					console.log('Queue insert failed');
+				}
+				catch (err) {
+					// console.log('Queue insert failed');
 					console.log(err);
 
 					reject(err);
 				}
 
-				console.log('Finished upload at ' + endTime.toString());
-				console.log('Total upload time: ' + endTime.diff(startTime, 'seconds') + ' seconds');
+				// console.log('Finished upload at ' + endTime.toString());
+				// console.log('Total upload time: ' + endTime.diff(startTime, 'seconds') + ' seconds');
 
 				resolve();
 			}
